First try with names.py
I am getting a little stumped on this one. I am pretty sure this is O(log n), but
the time complexity is 32 seconds. I tried to check the names as they were being inserted
to make it faster, but that fails because their is duplicates inside their own names file.
Need to keep thinking about it.